kernel: Guard TX descriptor with a spin lock
Since the receive reset mechanism may need to report a transmit
completion, we cannot rely solely on the transmit datapath behaviour
to eliminate the requirement for transmit locks.

Guard the transmit data structures with a spin lock.  The lock is
released prior to reporting a transmit completion to the IEEE 802.15.4
core, since reporting a completion may plausibly immediately trigger
the next transmission.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>