kernel: Check to see if hardware completely ignored the TXSTRT instruction
The internal state machines seem occasionally not to bother to notice
unimportant instructions such as "transmit packet".  Read back from
the SYS_CTRL register to make sure that the instruction has been seen.
If it has not, then immediately complete the packet to avoid a lockup
of the transmit datapath (since no TXFRS will ever be generated).

Also move the preceding TRXOFF before the write to TX_FCTRL, to allow
extra time for the TRXOFF to be noticed.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>