kernel: Resubmit the transmit instruction to work around hardware bugs
The DW1000 hardware occasionally decides to ignore the TXSTRT
instruction for no discernible reason.  We currently detect this
condition and immediately complete the transmission (thereby producing
an effect equivalent to a dropped network packet).

This causes problems for linuxptp, which will react to a missing
transmit timestamp by transitioning to a failure state from which it
will take several seconds to recover.

Work around the hardware bug by retrying the transmit instruction
several times if needed.  To minimise the amount of extra code
required, we reuse the original data SPI message.  We set the TX
buffer transfer set length to zero in order to reduce the amount of
data resent over the SPI bus.

Tested-by: Petri Mattila <petri.mattila@unipart.io>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>