kernel: Configure DRX_TUNE4H based on configured transmit preamble length
The value of the DRX_TUNE4H register is supposed to reflect the
expected receive preamble length.  This is not a well-defined concept,
since different transmitters on the same network may choose different
preamble lengths.  However, we need to specify some value for the
register: default to assuming that other transmitters will use the
same preamble length.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>