rpi: Replace DWM1000 module with discrete components and SMA connector
We want a board with a SMA connector, so we can't use the DWM1000 module
any more. This version of the PCB schematic implements more or less the same
functionality but with a SMA connector on the board, using discrete
components instead of the module.

It contains options for using an LDO or not for VBATT, using a TCXO or a
crystal, and using a DC-DC converter for the on-chip LDOs, or not.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>