rpi: B3 PCB with moved antenna, test pads and double sided connector
The DWM1000 is moved to the left in order to allow it to not overlap
with the ground plane when stacked on top of a PoE HAT. This means that
this board can no longer be called a HAT.

New test pads are provided to facilitate test and calibration measurements.

The connector for the Raspberry Pi can now be mounted on either side
of the board.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>