kernel: Use correct function to locate device tree calibration node
The "decawave,calib" node exists as a child of the dw1000 device's own
node, and so can be located using of_get_child_by_name().

The current code seems to accidentally use of_find_node_by_name(),
which is intended instead for iterating over a device tree (ignoring
the parent-child hierarchy).  This would usually happen to find the
correct node by pure luck, since there would be only a single
"decawave,calib" node in the tree and it would most likely directly
follow the dw1000 node.

of_find_node_by_name() will call of_node_put() on its argument, since
it is intended to be used in a pattern such as:

  while ((np = of_find_node_by_name(np, "name"))) { ... }

An unfortunate side effect of this accidental usage is therefore that
the reference count of the dw1000 node would be decremented by each
call to dw1000_load_profiles(), leading to a reference count underflow
and a subsequent kernel panic after four iterations of a driver
load/unload cycle.

Fix by using of_get_child_by_name(), which seems to match the expected
structure of the device tree.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>